Add get_gva_combine function
As an interim solution, add a function that allows the user to download the latest version of the gva_combine.Rmd notebook. Combining the data in a notebook with inline testing may be more appropriate than using the built in function -- or at least will allow easy edits if the function breaks due to changes in the incoming spreadsheets.